Add Hero component with Swiper functionality
This commit introduces a new Hero component that utilizes the Swiper
library to create a carousel feature. The Hero component is imported
and used in the Home page, replacing the previous MaxWidthWrapper
section. Swiper's CSS and Navigation modules are also imported to
support the carousel's styling and functionality. Additionally, new
background images for the carousel have been added to the project's
assets, and the Swiper library has been added to the package.json
dependencies. Corresponding styles for the Swiper component are
defined in a new styles.css file within the components/index directory.