Add country filter to getAllJobs action
Extend the getAllJobs function to support filtering by a list of countries. The function now accepts an optional 'country' parameter, which can be a comma-separated list of countries. This list is then used to construct a condition for the database query, allowing for more granular job search results based on the specified countries. The existing location-based filtering logic has been adjusted to accommodate this new feature.