Refactor UserMenuButton component to conditionally render dashboard link based on user role
The UserMenuButton component has been refactored to conditionally render the dashboard link based on the user's role. If the user is an admin, the link will direct to the admin dashboard, otherwise it will direct to the regular dashboard. This improves the user experience by providing relevant links based on their role.