Add Service component to admin service list
This commit adds the Service component to the admin service list. The component
displays the title and description of a service and provides options to view and
delete the service. The delete option is implemented using a ConfirmModal
component and a DELETE request to the server. The component also uses the
useToast hook to display success and error messages.