feat: Add service model and related API endpoints
This commit adds a new `Service` model to the Prisma schema and implements the necessary API endpoints for creating, updating, and deleting services. The `getAllServices` function is also added to retrieve all services from the database. Additionally, a new `TitleForm` component is created to allow editing of service titles in the admin dashboard. Some minor changes are also made to the `ServiceList` and `Service` components to use the new `slug` field instead of `id`. Finally, a new `getLatestFileMetaData` function is added to retrieve metadata for the latest file uploaded to Google Cloud Storage.