Add employment details route and comment out personal details form
This commit adds a new employment details route to the job seeker profile API. It also comments out the personal details form in the job seeker CV steps wrapper component. Additionally, it updates the employment details form to include a description field and fixes the component name of the skills form.