Add job type field to job creation form and model
This commit adds a new field for job type to the job creation form in the employer dashboard. It also adds the job type field to the Job model in the Prisma schema, with a default value of "NORMAL". Additionally, a new enum type "JOBTYPE" is created in the Prisma migration file.