Refactor personal details form and add scroll area component
This commit refactors the personal details form in the job seeker dashboard by adding a useEffect hook to fetch the user's personal details from the server and pre-populate the form fields. The date of birth field is now set to undefined by default. Additionally, a new ScrollArea component has been added to the UI library and is used in the ProfileCard component. The package.json and pnpm-lock.yaml files have been updated to include the new @radix-ui/react-scroll-area dependency. The eslint-plugin-import dependency has also been updated to version 2.29.1.