feat: Add application status to database schema and implement status change functionality
This commit adds a new enum type `ApplicationStatus` to the Prisma schema and updates the `Application` model to include a `status` field with a default value of `PENDING`. The commit also includes a new migration file to update the database schema.

In addition, the commit implements the functionality to change the status of a job application using a new API endpoint and a new form component. The commit also includes updates to the employer dashboard to display the application status and allow employers to change the status of an application.